+++ src/moonlink/src/storage/mooncake_table/snapshot_maintenance.rs
@@ -6,7 +6,29 @@ use crate::storage::mooncake_table::snapshot::SnapshotTableState;
 use crate::storage::mooncake_table::{
     DataCompactionPayload, FileIndiceMergePayload, MaintenanceOption, SnapshotTask,
 };
-use crate::storage::storage_utils::{TableId, TableUniqueFileId};
+use crate::storage::storage_utils::{ProcessedDeletionRecord, TableId, TableUniqueFileId};
+
+/// Remap single record location after compaction.
+/// Return if remap succeeds.
+fn remap_record_location_after_compaction(
+    deletion_log: &mut ProcessedDeletionRecord,
+    task: &mut SnapshotTask,
+) -> bool {
+    if task.data_compaction_result.is_empty() {
+        return false;
+    }
+
+    let old_record_location = &deletion_log.pos;
+    let remapped_data_files_after_compaction = &mut task.data_compaction_result;
+    let new_record_location = remapped_data_files_after_compaction
+        .remapped_data_files
+        .remove(old_record_location);
+    if new_record_location.is_none() {
+        return false;
+    }
+    deletion_log.pos = new_record_location.unwrap().record_location;
+    true
+}
 
 impl SnapshotTableState {
     /// ===============================
@@ -223,5 +245,52 @@ impl SnapshotTableState {
             task.index_merge_result.new_file_indices.clone(),
         )
         .await
+    }
+
+    /// Data compaction might delete existing persisted files, which invalidates record locations and requires a record location remap.
+    pub(super) fn remap_and_prune_deletion_logs_after_compaction(
+        &mut self,
+        task: &mut SnapshotTask,
+    ) {
+        // No need to prune and remap if no compaction happening.
+        if task.data_compaction_result.is_empty() {
+            return;
+        }
+
+        // Remap and prune committed deletion log.
+        let mut new_committed_deletion_log = vec![];
+        let old_committed_deletion_log = std::mem::take(&mut self.committed_deletion_log);
+        for mut cur_deletion_log in old_committed_deletion_log.into_iter() {
+            if let Some(file_id) = cur_deletion_log.get_file_id() {
+                // Case-1: the deletion log doesn't indicate a compacted data file.
+                if !task
+                    .data_compaction_result
+                    .old_data_files
+                    .contains(&file_id)
+                {
+                    new_committed_deletion_log.push(cur_deletion_log);
+                    continue;
+                }
+                // Case-2: the deletion log exists in the compacted new data file, perform a remap.
+                let remap_succ =
+                    remap_record_location_after_compaction(&mut cur_deletion_log, task);
+                if remap_succ {
+                    new_committed_deletion_log.push(cur_deletion_log);
+                    continue;
+                }
+                // Case-3: the deletion log doesn't exist in the compacted new file, directly remove it.
+            } else {
+                new_committed_deletion_log.push(cur_deletion_log);
+            }
+        }
+        self.committed_deletion_log = new_committed_deletion_log;
+
+        // Remap uncommitted deletion log.
+        for cur_deletion_log in &mut self.uncommitted_deletion_log {
+            if cur_deletion_log.is_none() {
+                continue;
+            }
+            remap_record_location_after_compaction(cur_deletion_log.as_mut().unwrap(), task);
+        }
     }
 }
